+++ packages/compiler/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @vue-vine/compiler
 
+## 1.7.10
+
+### Patch Changes
+
+- Refactor language service to using TS server request forwarding instead of WebSocket pipeline
+
+## 1.7.9
+
+### Patch Changes
+
+- Improve language service stability.
+
+## 1.7.8
+
+### Patch Changes
+
+- Fix language service globalTypes and bump to volar 2.4.20
+
 ## 1.7.7
 
 ### Patch Changes